Migrated validation logic unit tests
